% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantrines-package.R
\docType{package}
\name{quantrines-package}
\alias{quantrines}
\alias{quantrines-package}
\title{quantrines: Examples of quantile regression engines}
\description{
Provides a few quantile regression engines for use with a feature branch of the parsnip package.
}
\author{
\strong{Maintainer}: Daniel McDonald \email{daniel@stat.ubc.ca}

}
\keyword{internal}
