% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_interval_score.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(x, actual, na.rm = c("propogate", "drop", "fail"), ...)
}
\arguments{
\item{x}{A vector of class \link[=vec_quantiles]{vctrs_quantiles}.}

\item{actual}{double. Actual value(s)}

\item{na.rm}{character. How are NA predictions handled.}

\item{...}{not used}
}
\value{
a vector of nonnegative scores.
}
\description{
Weighted interval score (WIS), a well-known quantile-based
approximation of the commonly-used continuous ranked probability score
(CRPS). WIS is a proper score, and can be thought of as a distributional
generalization of absolute error. For example, see \href{https://arxiv.org/abs/2005.12881}{Bracher et al. (2020)} for discussion in the context
of COVID-19 forecasting.
}
\examples{
quantile_levels <- c(.2, .4, .6, .8)
predq_1 <- 1:4 #
predq_2 <- 8:11
vq <- vec_quantiles(list(predq_1, predq_2), quantile_levels)
actual <- c(3.3, 7.1)
weighted_interval_score(vq, actual)
weighted_interval_score(vq, 5.1)
}
